# Reduce encoding allocations