--- a/types/block.go
+++ b/types/block.go
@@ -113,13 +113,14 @@ func (b Block) ID() BlockID {
 func (b Block) MerkleRoot() crypto.Hash {
 	tree := crypto.NewTree()
 	var buf bytes.Buffer
+	e := encoder(&buf)
 	for _, payout := range b.MinerPayouts {
-		payout.MarshalSia(&buf)
+		payout.MarshalSia(e)
 		tree.Push(buf.Bytes())
 		buf.Reset()
 	}
 	for _, txn := range b.Transactions {
-		txn.MarshalSia(&buf)
+		txn.MarshalSia(e)
 		tree.Push(buf.Bytes())
 		buf.Reset()
 	}