@lukechampine 2017-09-07
neat -- turns out the stdlib jpeg encoder uses a similar strategy: https://golang.org/src/image/jpeg/writer.go#L216