@DavidVorick 2017-09-12
Did you do a full sync to double check that you didn't break any of the encoding?